rename Control class to Robot_Control - fixes #53
